refactor/client: zod user validation is now in its own function in a new utils folder
