+++ client/src/utils/generatePassword.ts
